refactor(core): centralize status next actions
